refactor(plugins/clients/tcp): move tcpConfig and add it as separate lib
