+++ src/plugins/clients/connectionmanager/private/connectionmanager.c
@@ -16,7 +16,6 @@
 #include "tcp_clientauthorization/clientauthorization.h"
 #include "tcp_config/config.h"
 
-
 static inline safuResultE_t elosUnixConfigGetSocketAddress(UNUSED elosPlugin_t const *plugin, UNUSED struct sockaddr *addr) {
     return SAFU_RESULT_FAILED;
 }
