+++ test/utest/plugins/clients/tcp/elosConfigGetElosdConnectionLimit/elosConfigGetElosdConnectionLimit_utest.h
@@ -4,7 +4,8 @@
 #include <cmocka_extensions/cmocka_extensions.h>
 #include <safu/mock_safu.h>
 #include <samconf/mock_samconf.h>
-#include <tcpConfig.h>
+
+#include "tcp_config/config.h"
 
 TEST_CASE_FUNC_PROTOTYPES(elosTestElosConfigGetElosdConnectionLimitSuccessFromEnv)
 TEST_CASE_FUNC_PROTOTYPES(elosTestElosConfigGetElosdConnectionLimitSuccessFromDefault)
