+++ test/utest/plugins/clients/tcp/elosConfigGetElosdInterface/elosConfigGetElosdInterface_utest.h
@@ -5,7 +5,8 @@
 #include <cmocka_extensions/cmocka_extensions.h>
 #include <safu/mock_safu.h>
 #include <samconf/mock_samconf.h>
-#include <tcpConfig.h>
+
+#include "tcp_config/config.h"
 
 TEST_CASE_FUNC_PROTOTYPES(elosTestElosConfigGetElosdInterfaceSuccessFromEnv)
 TEST_CASE_FUNC_PROTOTYPES(elosTestElosConfigGetElosdInterfaceSuccessFromDefault)
