+++ test/utest/plugins/clients/tcp/elosConfigGetElosdPort/elosConfigGetElosdPort_utest.h
@@ -5,7 +5,8 @@
 #include <cmocka_extensions/cmocka_extensions.h>
 #include <safu/mock_safu.h>
 #include <samconf/mock_samconf.h>
-#include <tcpConfig.h>
+
+#include "tcp_config/config.h"
 
 samconfConfig_t elosGetMockConfig();
